Embed Daily editing and viewing in Task and Resource pages
A Daily linked to a Task or Resource now lives inside that parent's
detail page: a `DailySection` renders the view panel inline and opens
an `Edit Daily` dialog in place. Direct hits on `/dailies/$id` and
`/dailies/$id/edit` for a linked Daily redirect to the parent with
search params (`dailySection`, `dailyMode`) that auto-open the embedded
section. Standalone Dailies (no Task, no Resource) keep their dedicated
edit page.

Extracts the existing form + view JSX into reusable `DailyForm`,
`DailyDetailsPanel`, `DailyEditDialog`, and `DailySection` components,
plus a `useRedirectLinkedDaily` hook for the redirect logic.
