+++ packages/client/src/components/dailies/index.ts
@@ -4,8 +4,12 @@ export { DailiesViewModeToggle } from "./DailiesViewModeToggle";
 export { DailyCommentPopover } from "./DailyCommentPopover";
 export { DailyCompletionsManager } from "./DailyCompletionsManager";
 export { DailyCourseIndicator } from "./DailyCourseIndicator";
+export { DailyDetailsPanel } from "./DailyDetailsPanel";
+export { DailyEditDialog } from "./DailyEditDialog";
+export { DailyForm } from "./DailyForm";
 export { DailyLocationCell } from "./DailyLocationCell";
 export { DailyProgressCell } from "./DailyProgressCell";
+export { DailySection } from "./DailySection";
 export { DailyTaskIndicator } from "./DailyTaskIndicator";
 export { DailyRecentDaysStrip } from "./DailyRecentDaysStrip";
 export { DailyStatusCircle } from "./DailyStatusCircle";
